+++ src/components/utilities/Defense.vue
@@ -43,6 +43,8 @@
                     <span class="defense-info__level-value">{{ defenseLevel }}</span>
                     <button class="btn btn-link" @click="defenseLevel++" :disabled="defenseLevel===5"><IconChevronUp /></button>
                   </div>
+
+                  <span class="badge rounded-pill bg-success du-badge">{{ defense.defenseData?.defenseUnits }} DU</span>
                 </span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Tooltip DPS:</strong> {{ Math.round(tooltipDps).toLocaleString('en-US') }}</span>
                 <span v-if="isDev" class="w-100 defense-info__header-stats__stat"><strong>Attack damage:</strong> {{ Math.round(attackDamage).toLocaleString('en-US') }}</span>
@@ -60,6 +62,8 @@
                     <span class="defense-info__level-value">{{ defenseLevel }}</span>
                     <button class="btn btn-link" @click="defenseLevel++" :disabled="defenseLevel===5"><IconChevronUp /></button>
                   </div>
+
+                  <span class="badge rounded-pill bg-success du-badge">{{ defense.defenseData.defenseUnits }} DU</span>
                 </span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Defense Power bonus:</strong> {{ Math.round(defensePower / 10) }}</span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Crit damage bonus:</strong> {{ (criticalDamage * 100 / 4).toFixed(2) }}%</span>
@@ -282,10 +286,12 @@ onMounted((): void => {
   font-size: larger;
   display: flex;
   flex-direction: column;
-  justify-content: center;
 }
 .defense-info__header-stats__stat {
   font-size: 12pt;
+}
+.du-badge {
+  margin-left: auto
 }
 
 .defense-dps {