+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/HEAD)
 
+* Added DU calculator to defense setups
 * Fixed a bug in the UI causing collapse to snap instead of transition
 
 [Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.1.8...HEAD)